requirements.txt
# Tubi Radar - Multi-Agent Competitive Intelligence System
# Python 3.9+

# LLM and Agents
langchain>=0.3.0
langchain-openai>=0.3.0
langgraph>=0.2.0
openai>=1.0.0

# Vector Store
chromadb>=0.5.0

# Database
sqlalchemy>=2.0.0

# RSS and HTTP
feedparser>=6.0.0
httpx>=0.27.0

# Configuration and Templating
pyyaml>=6.0
jinja2>=3.1.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Environment
python-dotenv>=1.0.0

# Utilities
beautifulsoup4>=4.12.0
lxml>=5.0.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0

# Dashboard
fastapi>=0.109.0
uvicorn>=0.27.0
jinja2>=3.1.0

